+++ tools/mhmake/src/md5.h
@@ -44,7 +44,7 @@ md5_context;
 typedef uint32 md5_val[4];
 
 void md5_starts( md5_context *ctx );
-void md5_update( md5_context *ctx, uint8 *input, uint32 length );
+void md5_update( md5_context *ctx, uint8 *input, size_t length );
 void md5_finish( md5_context *ctx, uint8 digest[16] );
 uint32 *md5_finishbin( md5_context *ctx);
 uint32 md5_finish32( md5_context *ctx);
